requirements.txt
crispy-forms-materialize==0.1
Django==1.8
django-allauth==0.29.0
django-crispy-forms==1.6.1
django-filter==1.0.1
django-materialize-css==0.0.1
django-nested-admin==3.0.12
djangorestframework==3.5.3
Markdown==2.6.7
oauthlib==2.0.1
Pillow==3.4.2
pkg-resources==0.0.0
python-monkey-business==1.0.0
python-openid==2.2.5
requests==2.12.4
requests-oauthlib==0.7.0
six==1.10.0
